+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 ### Changed
 - Change emailVerified User field with UserAttributeMappers (conditional trust email). [EOSC-KC-70](https://github.com/eosc-kc/keycloak/issues/70)
+- Support for configuring claims supported in Keycloak OP metadata [RCIAM-899](https://jira.argo.grnet.gr/browse/RCIAM-899)
 - Add is required configuration option for UserAttributeMapper and AttributeToRoleMapper 
 
 ### Fixed